# Conflicts:
#	.gradle/7.6.1/checksums/checksums.lock
#	.gradle/7.6.1/checksums/md5-checksums.bin
#	.gradle/7.6.1/checksums/sha1-checksums.bin
#	.idea/modules.xml
#	src/main/java/com/team/tesbro/Academy/AcademyController.java
#	src/main/resources/application.properties
#	src/main/resources/templates/academy_detail.html

+++ src/main/java/com/team/tesbro/User/UserRepository.java
@@ -4,6 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<SiteUser, Long> {
+public interface UserRepository extends JpaRepository<SiteUser, Integer> {
     Optional<SiteUser> findByusername(String username);
+
 }
